feat: Enhance user interactions and post management
- Updated Users bean to use Long for ID and added JsonProperty for password.
- Improved HomeController with a comment for future homepage implementation.
- Enhanced PostController with like/unlike functionality and updated ID types to Long.
- Modified UserController to handle Long ID for follow/unfollow actions.
- Added CORS configuration for development.
- Updated DevSecurityConfig to set Long ID for dummy user.
- Refactored FollowService to use Long for follower and followed IDs.
- Enhanced PostService with like/unlike methods and updated ID types to Long.
- Introduced InteractionType enum for post interactions.
- Created PostInteraction entity to track user interactions with posts.
- Added PostInteractionRepo for database operations related to post interactions.
- Updated frontend Card model to include Author interface.
- Enhanced recipe card component with like/dislike functionality and animations.
- Improved styling for recipe cards and added action buttons for interactions.
- Created InteractionService for handling post interactions in the frontend.
- Added workspace configuration for easier development setup.

+++ tastee-backend/src/main/java/com/tastee/tastee_backend/beans/Follow.java
@@ -9,24 +9,24 @@ import jakarta.persistence.IdClass;
 public class Follow {
 
     @Id
-    private int followerId;
+    private Long followerId;
 
     @Id
-    private int followedId;
+    private Long followedId;
 
-    public int getFollowerId() {
+    public Long getFollowerId() {
         return followerId;
     }
 
-    public void setFollowerId(int followerId) {
+    public void setFollowerId(Long followerId) {
         this.followerId = followerId;
     }
 
-    public int getFollowedId() {
+    public Long getFollowedId() {
         return followedId;
     }
 
-    public void setFollowedId(int followedId) {
+    public void setFollowedId(Long followedId) {
         this.followedId = followedId;
     }
